+++ CHANGELOG
@@ -1,0 +1,8 @@
+* v.0.0.4
+И эта версия тоже с багами, но уже лучше =)
+* v.0.0.3
+Третья тоже с багами
+* v.0.0.2
+Вторая версия с багами
+* v.0.0.1
+Первая версия увидевшая свет
