+++ README.md
@@ -1,8 +1,10 @@
 # О проекте
-Управление компьютером через веб интерфейс на [PHP](http://php.net/), например: сделать звук тише, громче или выключить. Безопастность пока хромает, выпускать в интернет не надо. Работает на локальном веб сервере в составе OpenServer. Управляет компьютером утилита [nirCmd](http://nircmd.nirsoft.net/).
+Управление компьютером через веб интерфейс на [PHP](http://php.net/), например: сделать звук тише, громче или выключить.
+Безопастность пока хромает, выпускать в интернет не надо. Работает на локальном веб сервере в составе OpenServer.
+Управляет компьютеромутилита[nirCmd](http://nircmd.nirsoft.net/).
 
 
-# Реализованы следующие функции
+# Доступен функционал
 * `standby` - спящий режим
 * `hibernate`- гибернация
 * `reboot` - перезагрузится
@@ -17,7 +19,7 @@
 * `unmute` - включить звук
 * `triggerMute` - включить/отключить звук по триггеру
 * `volUp` - громкость больше
-* `volDown` - громкость меньше 
+* `volDown` - громкость меньше
 
 
 # Attention machine translation
