+++ app/autoload.php
@@ -1,0 +1,10 @@
+<?php
+
+/**
+ * Автозагрузка классов
+ */
+function __autoload($className) {
+    require PATH . '/class/' . $className . '.php';
+}
+
+return new Application;
