+++ app/class/ActionRoute.class.php
@@ -1,0 +1,5 @@
+<?php
+interface ActionRoute {
+    function route($arr); //проверка данных
+    function execute($c); //исполнение команды
+}
