+++ app/class/Message.class.php
@@ -1,0 +1,13 @@
+<?php
+
+class Message {
+    /**
+     * Отправка сообщения
+     * @method sent
+     * @param string $messageText текст сообщения
+     * @return
+     */
+    public static function sent($messageText) {
+        echo json_encode($m);
+    }
+}
