+++ app/config/config.php
@@ -1,0 +1,11 @@
+<?php
+
+/**
+ * Конфиг для настройки
+ */
+return [
+   'filename'     => PATH . '/nircmd/nircmd.exe',
+   'login'        => 'test',
+   'psw'          => 'test',
+   'debug'        => true
+];
