--- class/ActionRoute.class.php
@@ -1,5 +1,0 @@
-<?php
-interface ActionRoute {
-    function route($arr); //проверка данных
-    function execute($c); //исполнение команды
-}
