--- class/Message.class.php
@@ -1,11 +1,0 @@
-class Message {
-    /**
-     * Отправка сообщения
-     * @method sent
-     * @param string $messageText текст сообщения
-     * @return
-     */
-    public static function sent($messageText) {
-        echo json_encode($m);
-    }
-}
